Refactor actor runtime callback scheduler to enforce durable-only semantics
- Updated the `IActorRuntimeCallbackScheduler` interface and its implementations to ensure all callbacks are scheduled with durable semantics, removing inline and dedicated options.
- Enhanced the `EventHandlerContext` and `GAgentBase` classes to utilize the new durable scheduling methods, improving clarity and consistency in callback handling.
- Removed legacy inline callback scheduling components to streamline the architecture and prevent misuse of transient state.
- Documented architectural changes and updated dependency injection configurations to reflect the new durable callback model.
- Improved metadata validation for scheduled callbacks, ensuring robust handling of callback leases and enhancing overall system reliability.
