--- src/Aevatar.Foundation.Runtime.Implementations.Orleans/Grains/IRuntimeActorInlineCallbackSchedulerBindingAccessor.cs
@@ -1,8 +1,0 @@
-namespace Aevatar.Foundation.Runtime.Implementations.Orleans.Grains;
-
-public interface IRuntimeActorInlineCallbackSchedulerBindingAccessor
-{
-    IRuntimeActorInlineCallbackScheduler? Current { get; }
-
-    IDisposable Bind(IRuntimeActorInlineCallbackScheduler scheduler);
-}
