+++ test/Aevatar.Workflow.Extensions.Maker.Tests/MakerRecursiveModuleCoverageTests.cs
@@ -222,7 +222,7 @@ public class MakerRecursiveModuleCoverageTests
             return Task.CompletedTask;
         }
 
-        public Task<RuntimeCallbackLease> ScheduleSelfTimeoutAsync(
+        public Task<RuntimeCallbackLease> ScheduleSelfDurableTimeoutAsync(
             string callbackId,
             TimeSpan dueTime,
             IMessage evt,
@@ -237,7 +237,7 @@ public class MakerRecursiveModuleCoverageTests
             throw new NotSupportedException("This test context does not support scheduling.");
         }
 
-        public Task<RuntimeCallbackLease> ScheduleSelfTimerAsync(
+        public Task<RuntimeCallbackLease> ScheduleSelfDurableTimerAsync(
             string callbackId,
             TimeSpan dueTime,
             TimeSpan period,
@@ -254,7 +254,7 @@ public class MakerRecursiveModuleCoverageTests
             throw new NotSupportedException("This test context does not support scheduling.");
         }
 
-        public Task CancelScheduledCallbackAsync(
+        public Task CancelDurableCallbackAsync(
             RuntimeCallbackLease lease,
             CancellationToken ct = default)
         {
