+++ test/Aevatar.Workflow.Extensions.Maker.Tests/MakerVoteModuleCoverageTests.cs
@@ -220,7 +220,7 @@ public class MakerVoteModuleCoverageTests
             return Task.CompletedTask;
         }
 
-        public Task<RuntimeCallbackLease> ScheduleSelfTimeoutAsync(
+        public Task<RuntimeCallbackLease> ScheduleSelfDurableTimeoutAsync(
             string callbackId,
             TimeSpan dueTime,
             IMessage evt,
@@ -235,7 +235,7 @@ public class MakerVoteModuleCoverageTests
             throw new NotSupportedException("This test context does not support scheduling.");
         }
 
-        public Task<RuntimeCallbackLease> ScheduleSelfTimerAsync(
+        public Task<RuntimeCallbackLease> ScheduleSelfDurableTimerAsync(
             string callbackId,
             TimeSpan dueTime,
             TimeSpan period,
@@ -252,7 +252,7 @@ public class MakerVoteModuleCoverageTests
             throw new NotSupportedException("This test context does not support scheduling.");
         }
 
-        public Task CancelScheduledCallbackAsync(
+        public Task CancelDurableCallbackAsync(
             RuntimeCallbackLease lease,
             CancellationToken ct = default)
         {
